requirements.txt
# Core dependencies
python-dotenv==1.0.0  # For loading environment variables from .env files
pydantic==2.4.2       # For data validation and settings management
requests==2.31.0      # For API communication

# Logging
loguru==0.7.0         # Enhanced logging capabilities

# Testing
pytest==7.4.0         # Testing framework
pytest-mock==3.11.1   # Mocking library for pytest
coverage==7.3.0       # Code coverage measurement

# Development tools
black==23.7.0         # Code formatting
flake8==6.1.0         # Linting
mypy==1.5.1           # Static type checking
isort==5.12.0         # Import sorting

# Documentation
sphinx==7.1.2         # Documentation generator
sphinx-rtd-theme==1.3.0  # Read the Docs theme for Sphinx
